requirements.txt
pandas>=1.5.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0
pytest>=6.0.0
jupyter>=1.0.0
openpyxl>=3.0.0
sqlalchemy>=1.4.0
